fixes missing batch problem

+++ cdvae/pl_data/dataset.py
@@ -46,7 +46,7 @@ class CrystDataset(Dataset):
         data = Data(
             coords=torch.Tensor(data_dict["coords"]),
             atom_types=torch.Tensor(data_dict["elements"]),
-            num_atoms=torch.Tensor([data_dict["num_atoms"]]),
+            num_atoms=torch.Tensor([data_dict["num_atoms"]]).long(),
             y=prop
         )
